Automatic profile loading on reception of SIGUSR1
The ~/.screenlayout directory will be searched for profiles in JSON format. If a found profile contains outputs that exactly match the current configuration, the profile will be loaded.

The number and EDID of monitors in the saved profile must be the same as those currently connected to the PC (no more, no less).